refactor(workspace): relocate concurrent-search tests; self-contained test doubles
Move the white-box tests for ConcurrentSearch/searchPipeline/checkBinary/formatMatch into internal/pkg/concurrentsearch, replacing the infra and toolstest test doubles (NewWorkspacePolicy, isPathSafeErrorSM) with local domain-only mocks (mockPolicy, mockSPError) so the package's merged closure stays {persistence,security,services} — self-justifying under ADR-056, no whitelist entry. Delete the four dead workspace symbols; repoint the benchmark to the new package.
